Move receipt printing to shared function
