+++ WooCommerce/Classes/ViewModels/CardPresentPayments/PaymentCaptureOrchestrator.swift
@@ -81,21 +81,6 @@ final class PaymentCaptureOrchestrator {
         ServiceLocator.stores.dispatch(action)
     }
 
-    func printReceipt(for order: Order, params: CardPresentReceiptParameters) {
-        let action = ReceiptAction.print(order: order, parameters: params) { (result) in
-            switch result {
-            case .success:
-                ServiceLocator.analytics.track(.receiptPrintSuccess)
-            case .cancel:
-                ServiceLocator.analytics.track(.receiptPrintCanceled)
-            case .failure(let error):
-                ServiceLocator.analytics.track(.receiptPrintFailed, withError: error)
-            }
-        }
-
-        ServiceLocator.stores.dispatch(action)
-    }
-
     func emailReceipt(for order: Order, params: CardPresentReceiptParameters, onContent: @escaping (String) -> Void) {
         let action = ReceiptAction.generateContent(order: order, parameters: params) { emailContent in
             onContent(emailContent)
