+++ WooCommerce/Classes/ViewModels/CardPresentPayments/ReceiptPrintingCoordinator.swift
@@ -1,0 +1,21 @@
+import Foundation
+import Yosemite
+
+struct ReceiptPrintingCoordinator {
+    static func printReceipt(for order: Order, params: CardPresentReceiptParameters) {
+        ServiceLocator.analytics.track(.receiptPrintTapped)
+
+        let action = ReceiptAction.print(order: order, parameters: params) { (result) in
+            switch result {
+            case .success:
+                ServiceLocator.analytics.track(.receiptPrintSuccess)
+            case .cancel:
+                ServiceLocator.analytics.track(.receiptPrintCanceled)
+            case .failure(let error):
+                ServiceLocator.analytics.track(.receiptPrintFailed, withError: error)
+            }
+        }
+
+        ServiceLocator.stores.dispatch(action)
+    }
+}
