+++ WooCommerce/Classes/ViewModels/CardPresentPayments/ReceiptViewModel.swift
@@ -29,17 +29,6 @@ final class ReceiptViewModel {
 
     /// Prints the receipt
     func printReceipt() {
-        ServiceLocator.analytics.track(.receiptPrintTapped)
-        let action = ReceiptAction.print(order: order, parameters: receipt) { (result) in
-            switch result {
-            case .success:
-                ServiceLocator.analytics.track(.receiptPrintSuccess)
-            case .cancel:
-                ServiceLocator.analytics.track(.receiptPrintCanceled)
-            case .failure(let error):
-                ServiceLocator.analytics.track(.receiptPrintFailed, withError: error)
-            }
-        }
-        ServiceLocator.stores.dispatch(action)
+        ReceiptPrintingCoordinator.printReceipt(for: order, params: receipt)
     }
 }
