+++ WooCommerce/Classes/ViewModels/Order Details/OrderDetailsViewModel.swift
@@ -550,8 +550,7 @@ extension OrderDetailsViewModel {
     }
 
     func printReceipt(params: CardPresentReceiptParameters) {
-        ServiceLocator.analytics.track(.receiptPrintTapped)
-        paymentOrchestrator.printReceipt(for: order, params: params)
+        ReceiptPrintingCoordinator.printReceipt(for: order, params: params)
     }
 
     func emailReceipt(params: CardPresentReceiptParameters, onContent: @escaping (String) -> Void) {
